+++ src/components/_common/qbrick-video/QbrickVideo.module.scss
