+++ src/components/_common/qbrick-video/QbrickVideo.tsx
@@ -1,0 +1,235 @@
+import React, { useCallback, useEffect, useId, useState } from 'react';
+import { AnalyticsEvents, logAmplitudeEvent } from 'utils/amplitude';
+import { Button, Detail, Label, Loader } from '@navikt/ds-react';
+import { getMediaUrl } from 'utils/urls';
+import {
+    findImageUrlFromVideoMeta,
+    findVideoDurationFromMeta,
+    getTimestampFromDuration,
+    buildVideoMeta,
+} from './videoHelpers';
+import { translator } from 'translations';
+import { usePageConfig } from 'store/hooks/usePageConfig';
+import { fetchJson } from 'srcCommon/fetch-utils';
+import Script from 'next/script';
+import { classNames } from 'utils/classnames';
+import { AlertBox } from 'components/_common/alert-box/AlertBox';
+import { logger } from 'srcCommon/logger';
+import { EditorHelp } from 'components/_editor-only/editor-help/EditorHelp';
+
+import style from './QbrickVideo.module.scss';
+
+const PLAYER_TIMEOUT_MS = 5000;
+const PLAYER_POLLING_RATE_MS = 50;
+
+export type VideoMeta = {
+    accountId: string;
+    mediaId: string;
+    title: string;
+    duration: number;
+    poster?: string;
+    language?: string;
+};
+
+type Props = {
+    videoMeta: VideoMeta;
+};
+
+export const QbrickVideo = ({ videoMeta }: Props) => {
+    const { language: contentLanguage, pageConfig } = usePageConfig();
+    const { editorView, pageId } = pageConfig;
+    const translations = translator('macroVideo', contentLanguage);
+
+    const [isPlayerLoading, setIsPlayerLoading] = useState(false);
+    const [isPlayerReady, setIsPlayerReady] = useState(false);
+    const [isPlayerError, setIsPlayerError] = useState(false);
+
+    const videoRef = React.useRef(null);
+
+    const widgetId = useId();
+
+    if (!videoMeta) {
+        return <EditorHelp type={'error'} text={'Ugyldig video macro'} />;
+    }
+
+    const {
+        accountId,
+        mediaId,
+        title,
+        duration,
+        poster,
+        language: videoLanguage,
+    } = videoMeta;
+
+    const getVideoMetaFromQbrick = useCallback(async () => {
+        const metaUrl = `https://video.qbrick.com/api/v1/public/accounts/${accountId}/medias/${mediaId}`;
+
+        try {
+            const result = await fetchJson(metaUrl);
+            if (!result) {
+                return;
+            }
+
+            const poster = findImageUrlFromVideoMeta(result);
+            const duration = findVideoDurationFromMeta(result);
+
+            setVideoMeta({ ...videoMeta, poster, duration });
+        } catch (e) {
+            logger.error(e);
+            setIsPlayerError(true);
+        }
+    }, [accountId, mediaId, videoMeta]);
+
+    const createAndStartPlayer = useCallback(
+        (timeLeft = PLAYER_TIMEOUT_MS) => {
+            if (timeLeft <= 0) {
+                logger.error('Failed to load QBrick player - Timed out');
+                setIsPlayerError(true);
+                return;
+            }
+
+            // Should be defined when the GoBrain init script has finished executing
+            // There doesn't seem to be an elegant way to consistently determine when
+            // this has happened, so we do some polling...
+            if (!window.GoBrain) {
+                setTimeout(
+                    () =>
+                        createAndStartPlayer(timeLeft - PLAYER_POLLING_RATE_MS),
+                    PLAYER_POLLING_RATE_MS
+                );
+                return;
+            }
+
+            const widgetExists = !!window.GoBrain.widgets(widgetId);
+            if (widgetExists) {
+                return;
+            }
+
+            if (!videoRef.current) {
+                return;
+            }
+
+            window.GoBrain.create(videoRef.current, {
+                config: `//video.qbrick.com/play2/api/v1/accounts/${accountId}/configurations/qbrick-player`,
+                data: `//video.qbrick.com/api/v1/public/accounts/${accountId}/medias/${mediaId}`,
+                language: videoLanguage,
+                autoplay: true,
+                widgetId,
+            }).on('ready', () => {
+                setIsPlayerReady(true);
+                setIsPlayerLoading(false);
+                logAmplitudeEvent(AnalyticsEvents.VIDEO_START, {
+                    tittel: title,
+                    varighet: duration,
+                    språk: videoLanguage,
+                });
+            });
+        },
+        [widgetId, accountId, mediaId, title, duration, videoLanguage]
+    );
+
+    const resetPlayerState = useCallback(() => {
+        setIsPlayerReady(false);
+        setIsPlayerError(false);
+
+        if (window.GoBrain) {
+            window.GoBrain.destroy(widgetId, true);
+        }
+    }, [widgetId]);
+
+    useEffect(() => {
+        // Whether the video is in new content or legacy, attempt
+        // to get the poster and duration if none is given in the content config.
+        if (!videoMeta.poster && !videoMeta.duration) {
+            getVideoMetaFromQbrick();
+        }
+    }, [videoMeta.poster, videoMeta.duration, getVideoMetaFromQbrick]);
+
+    useEffect(() => {
+        setVideoMeta(buildVideoMeta(config?.video, contentLanguage));
+        return resetPlayerState;
+    }, [pageId, resetPlayerState, config, contentLanguage]);
+
+    if (!accountId) {
+        return null;
+    }
+
+    const durationAsString = getTimestampFromDuration(duration);
+    const imageUrl = poster?.startsWith('http')
+        ? poster
+        : getMediaUrl(poster, !!editorView);
+
+    return (
+        <div className={style.wrapper}>
+            <Script
+                src={
+                    'https://play2.qbrick.com/qbrick-player/framework/GoBrain.min.js'
+                }
+                async={true}
+                onError={() => {
+                    setIsPlayerError(true);
+                }}
+            />
+            <Button
+                className={classNames(
+                    style.button,
+                    isPlayerReady && style.hidden
+                )}
+                variant={'tertiary'}
+                onClick={() => {
+                    if (editorView !== 'edit' && !isPlayerLoading) {
+                        setIsPlayerLoading(true);
+                        createAndStartPlayer();
+                    }
+                }}
+                icon={
+                    <div className={style.posterWrapper}>
+                        <img
+                            className={style.previewImage}
+                            src={imageUrl}
+                            alt={''}
+                        />
+                        <div className={style.playBadge}>
+                            {isPlayerLoading ? (
+                                <Loader className={style.playLoader} />
+                            ) : (
+                                <svg
+                                    className={style.playArrow}
+                                    focusable={'false'}
+                                    aria-hidden={'true'}
+                                    viewBox={'0 0 22 26'}
+                                >
+                                    <path fill={'#fff'} d={'M22 13 0 26V0Z'} />
+                                </svg>
+                            )}
+                        </div>
+                    </div>
+                }
+            >
+                <Label as={'p'} className={style.text}>
+                    {`${translations('playMovie')} ${title}`}
+                </Label>
+                {duration > 0 && (
+                    <Detail
+                        className={classNames(style.text, style.videoLength)}
+                    >
+                        {`${translations(
+                            'duration'
+                        )} ${durationAsString} ${translations('minutes')}`}
+                    </Detail>
+                )}
+            </Button>
+            {isPlayerError && (
+                <AlertBox variant={'error'}>{translations('error')}</AlertBox>
+            )}
+            <div
+                className={classNames(
+                    style.macroVideo,
+                    !isPlayerReady && style.hidden
+                )}
+                ref={videoRef}
+                title={title}
+            />
+        </div>
+    );
+};
