+++ src/components/pages/video-preview-page/VideoPreviewPage.module.scss
