+++ src/pages/helloworld.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Test = () => {
+    return <div>{'Hello world!'}</div>;
+};
+
+export default Test;
